+++ .phpstan-baseline.neon
@@ -781,7 +781,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method guidFromUri\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:guidFromUri instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:guidFromUri instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 2
@@ -790,7 +790,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -799,7 +799,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method guidFromUri\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:guidFromUri instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:guidFromUri instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -808,7 +808,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -817,7 +817,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -826,7 +826,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method guidFromUri\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:guidFromUri instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:guidFromUri instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -835,7 +835,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -844,7 +844,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -853,7 +853,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -862,7 +862,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
@@ -871,7 +871,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method guidFromUri\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:guidFromUri instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:guidFromUri instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 2
@@ -880,7 +880,7 @@ parameters:
 		-
 			message: '''
 				#^Call to deprecated method newURI\(\) of class Friendica\\Model\\Item\:
-				2026\.08 Use \\Friendica\\Item\\UriGenerator\:\:newURI instead$#
+				2026\.08 Use \\Friendica\\Post\\UriGenerator\:\:newURI instead$#
 			'''
 			identifier: staticMethod.deprecated
 			count: 1
